refactor: Renames property.py to schema.py
