+++ tests/unit_tests/diffs/test_core.py
@@ -6,7 +6,7 @@ from schemaver.changelog import ChangeLevel
 from schemaver.diffs.core import CoreField
 from schemaver.diffs.string import StringField
 from schemaver.diffs.numeric import NumericField
-from schemaver.property import InstanceType, Property
+from schemaver.schema import InstanceType, Property
 
 from tests.unit_tests.diffs.helpers import (
     assert_changes,
