+++ tests/unit_tests/diffs/test_numeric.py
@@ -8,7 +8,7 @@ from schemaver.changelog import ChangeLevel, Changelog
 from schemaver.diffs.array import ArrayField
 from schemaver.diffs.string import StringField
 from schemaver.diffs.numeric import NumericField
-from schemaver.property import InstanceType, Property
+from schemaver.schema import InstanceType, Property
 
 from tests.unit_tests.diffs.helpers import (
     assert_changes,
