+++ tests/unit_tests/diffs/test_string.py
@@ -5,7 +5,7 @@ import pytest
 from schemaver.changelog import ChangeLevel
 from schemaver.diffs.array import ArrayField
 from schemaver.diffs.string import StringField
-from schemaver.property import InstanceType, Property
+from schemaver.schema import InstanceType, Property
 
 from tests.unit_tests.diffs.helpers import (
     assert_changes,
